requirements.txt
torch
scikit-image
numpy
matplotlib
argparse
Pillow
streamlit